Introduce secp256k1 and ed25519 modules and improve docs
